requirements.txt
feedparser
google-generativeai
tweepy
